fix erlyweb_compile runtime error happening on OTP R15B+.
If a erlyweb module is not defining the before_compile hook, the
call to apply/3 that erlyweb_compile:try_func/4 performs returns an
error tuple with the reason `undef` (expected) and a list whose 1st
element is a 4 terms tuple on R15B and R16B. (The same might be
true for later OTPs). On R14B a 3 terms tuple is returned instead.